[cmd] Consolidate files
